reid_app/pyproject.toml
[project]
name = "frigate-reid"
version = "0.1.0"
description = "Frigate Person Re-Identification with OpenVINO and FastAPI"
authors = [
    {name = "Your Name", email = "you@example.com"},
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "paho-mqtt>=2.0.0",
    "requests>=2.31.0",
    "opencv-python-headless>=4.9.0",
    "openvino>=2023.3.0",
    "numpy>=1.26.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
